requirements.txt
anyio==4.5.2
attrs==19.3.0
Automat==0.8.0
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
colorama==0.4.3
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cymem==2.0.11
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
editdistpy==0.1.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.115.11
fasttext-wheel==0.9.2
Flask==3.0.3
h11==0.14.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib_metadata==8.5.0
incremental==16.10.1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
langcodes==3.4.1
language-selector==0.1
language_data==1.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
marisa-trie==1.2.1
MarkupSafe==2.1.5
more-itertools==4.2.0
murmurhash==1.0.12
netifaces==0.10.4
nltk==3.9.1
numpy==1.24.4
oauthlib==3.1.0
packaging==20.3
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.6.0
preshed==3.0.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.13.6
pydantic==1.10.21
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.10.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-debian==0.1.36+ubuntu1.1
PyYAML==5.3.1
rank-bm25==0.2.2
regex==2024.11.6
requests==2.22.0
requests-unixsocket==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
smart-open==6.4.0
sniffio==1.3.1
sos==4.7.2
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.10
starlette==0.44.0
supervisor==4.1.0
symspellpy==6.7.8
systemd-python==234
thinc==8.1.12
threadpoolctl==3.5.0
tqdm==4.67.1
Twisted==18.9.0
typer==0.7.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.33.0
wadllib==1.3.3
wasabi==1.1.3
Werkzeug==3.0.6
zipp==3.20.2
zope.interface==4.7.1
openai==0.28.0